requirements.txt
gspread
pandas
selenium
tqdm
bs4
requests
python-dotenv